Update types based on PR feedback
